marketing-agent/pyproject.toml
[project]
name = "marketing-agent"
version = "0.1.0"
description = "Universal AI marketing + SEO team powered by Claude Code Agent Teams"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
